+++ spring-cloud-starter-tencent-polaris-ratelimit/src/main/java/com/tencent/cloud/polaris/ratelimit/config/PolarisRateLimitProperties.java
@@ -48,6 +48,18 @@ public class PolarisRateLimitProperties {
 	 */
 	private long maxQueuingTime = 1000L;
 
+	private String limiterAddresses;
+
+	private long remoteTaskInterval = 30L;
+
+	String getLimiterAddresses() {
+		return limiterAddresses;
+	}
+
+	void setLimiterAddresses(String limiterAddresses) {
+		this.limiterAddresses = limiterAddresses;
+	}
+
 	public String getRejectRequestTips() {
 		return rejectRequestTips;
 	}
@@ -78,5 +90,13 @@ public class PolarisRateLimitProperties {
 
 	public void setMaxQueuingTime(long maxQueuingTime) {
 		this.maxQueuingTime = maxQueuingTime;
+	}
+
+	public long getRemoteTaskInterval() {
+		return remoteTaskInterval;
+	}
+
+	public void setRemoteTaskInterval(long remoteTaskInterval) {
+		this.remoteTaskInterval = remoteTaskInterval;
 	}
 }