+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - [feat: support config event and monitor address list.](https://github.com/Tencent/spring-cloud-tencent/pull/1692)
 - [feat:Modify the annotations.](https://github.com/Tencent/spring-cloud-tencent/pull/1693)
 - [fix: polaris.discovery.heartbeat.enabled not effective.](https://github.com/Tencent/spring-cloud-tencent/pull/1694)
+- [feat: support config ratelimit addresses and remote task interval.](https://github.com/Tencent/spring-cloud-tencent/pull/1679)